+++ src/test/documentation_examples.rs
@@ -1756,7 +1756,7 @@ async fn index_examples() -> Result<()> {
 }
 
 async fn change_streams_examples() -> Result<()> {
-    use crate::{action::FullDocumentType, runtime};
+    use crate::{options::FullDocumentType, runtime};
     use std::time::Duration;
 
     let client = TestClient::new().await;
